// sanity/types.ts
export type SanityImageCrop = {
  _type: "sanity.imageCrop";
  top: number;
  bottom: number;
  left: number;
  right: number;
};

export type SanityImageHotspot = {
  _type: "sanity.imageHotspot";
  x: number;
  y: number;
  height: number;
  width: number;
};
// ...
export declare const internalGroqTypeReferenceTo: unique symbol;
// Source: sanity/lib/queries.ts
// Variable: HEADERQUERY
// Query: *[_type == 'header'][0]
export type HEADERQUERYResult = {
  _id: string;
  _type: "header";
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  title?: string;
  logo?: {
    asset?: {
      _ref: string;
      _type: "reference";
      _weak?: boolean;
      [internalGroqTypeReferenceTo]?: "sanity.imageAsset";
    };
    media?: unknown;
    hotspot?: SanityImageHotspot;
    crop?: SanityImageCrop;
    altText?: string;
    _type: "image";
  };
  menu?: Array<{
    title?: string;
    slug?: string;
    hasSubMenu?: boolean;
    subMenus?: Array<{
      title?: string;
      slug?: string;
      _type: "subMenu";
      _key: string;
    }>;
    _type: "menuItem";
    _key: string;
  }>;
  cta?: string;
} | null;